requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Faker==37.1.0
fastapi==0.111.0
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic-settings==2.2.1
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
